Create footerContactUsLink.js
The footer of Social Cloud pages has a "Contact Us" link that sends an email to L1 and gets the PMR process started.  We are wondering if that link could be adjusted to send an email instead to a given organization's help desk.  This commit will test if we can use Customizer to complete the request change.